Refactor: Rename setup.py and improve uninstall cleanup
- Rename `ury/setup.py` to `ury/setup_customizations.py` to better reflect its purpose.
- Update import statements in `install.py` and `uninstall.py` to use the new module name.
- Enhance `uninstall()` in `uninstall.py` by ensuring custom fields and customizations are properly removed with added error handling.
