+++ ury/uninstall.py
@@ -1,7 +1,7 @@
 import click
 import frappe
 
-from ury.setup import before_uninstall as remove_custom_fields
+from ury.setup_customizations import before_uninstall as remove_custom_fields
 
 
 def before_uninstall():
@@ -13,6 +13,12 @@ def before_uninstall():
         print("Failed To Remove Customizations.")
 
 def uninstall():
+	try:
+		print("Removing customizations created by the Frappe URY app...")
+		remove_custom_fields()
+	except Exception as e:
+		print(f"Failed To Remove Customizations: {e}")
+
 	ROLES = ["URY Manager", "URY Captain", "URY Cashier"]
 
 	frappe.db.delete("Custom DocPerm",{"role": ["in", ROLES]})
